+++ apps/web/src/app/components/input/page.tsx
@@ -7,6 +7,7 @@ import { InputDisabled } from "./variants/disabled";
 import { InputErrorState } from "./variants/error-state";
 import { InputHelperText } from "./variants/helper-text";
 import { InputLabelHelper } from "./variants/label-helper";
+import { InputLabelPlacement } from "./variants/label-placement";
 import { InputOptional } from "./variants/optional";
 import { InputPlaceholder } from "./variants/placeholder";
 import { InputRequired } from "./variants/required";
@@ -43,6 +44,9 @@ const Input = () => {
       </VariantWrapper>
       <VariantWrapper heading="With Placeholder">
         <InputWithPlaceholder />
+      </VariantWrapper>
+      <VariantWrapper heading="Label Placement">
+        <InputLabelPlacement />
       </VariantWrapper>
       <VariantWrapper heading="WithoutLabel">
         <InputWithoutLabel />