+++ apps/web/src/app/components/input/variants/variants.tsx
@@ -4,16 +4,12 @@ export const InputVariants = () => {
   return (
     <div className="space-y-4">
       <Input
-        label="Outline"
-        variant="outline"
+        label="Outlined"
+        variant="outlined"
       />
       <Input
         label="Filled"
         variant="filled"
-      />
-      <Input
-        label="Transparent"
-        variant="transparent"
       />
     </div>
   );