+++ apps/web/tailwind.config.ts
@@ -19,20 +19,7 @@ const config: Config = {
   },
   plugins: [
     ...jamsrui({
-      // colors: {
-      //   dark: {
-      //     default: {
-      //       DEFAULT: "#fff",
-      //       foreground: "#000",
-      //     },
-      //   },
-      //   light: {
-      //     default: {
-      //       DEFAULT: "#000",
-      //       foreground: "#fff",
-      //     },
-      //   },
-      // },
+      
     }),
   ],
 };