+++ packages/components/autocomplete/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/autocomplete
 
+## 0.0.61
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/input@0.0.56
+
 ## 0.0.60
 
 ### Patch Changes